+++ tests/data/odoo/addons/addon2/__manifest__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+{
+    'author': 'ACSONE SA/NV',
+    'license': 'AGPL-3',
+    'depends': [
+        'base',
+    ],
+}